Add a button for force push alliances to TBA after selection
If Cheesy Arena does not have Internet when alliances are finalized,
there's no way to republish the alliances. So add a button to do so if
selection has already completed.

Also don't show a misleading "Start Alliance Selection" button if it's
just going to error out that alliances have already been finalized.