Add mode for setting AP networks slowly
This commit adds a configuration setting that will commit team network
configurations to the AP one at a time instead of all in one message.
This is how FIRST FMS configures the AP, it seems to be more reliable
due to a bug in OpenWRT when you commit the entire config at once.

Update test