Add FRC Nexus Integration to get team lineup
frc.nexus Integration to pull team lineups for practice matches and playoffs.

Caveats that need to be addressed.
js file uses static value for event code, this should be automated before merging.

For playoffs the scorekeeper must be logged into Nexus with scorekeeper permissions, this is because the lineups are considered privileged and are not in the public API.